Refactor local vault controllers into service.local
.interface_adapter to follow existing pattern in folder structure
