+++ src/main/java/service/local/interface_adapter/CreateLocalVaultController.java
@@ -1,4 +1,4 @@
-package presenters.controllers;
+package service.local.interface_adapter;
 
 import javax.swing.JFileChooser;
